# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: FishyKeys API
    description: The FishyKeys API for key management
    version: "1.0"
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /users:
        get:
            tags:
                - users
            summary: list users
            description: List all users
            operationId: users#list
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/User'
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
            schemes:
                - http
        post:
            tags:
                - users
            summary: create users
            description: Create a new user
            operationId: users#create
            parameters:
                - name: CreateRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/UsersCreateRequestBody'
                    required:
                        - username
                        - password
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/UsersCreateResponseBody'
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "409":
                    description: Conflict response.
                    schema:
                        type: string
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
            schemes:
                - http
    /users/{username}:
        delete:
            tags:
                - users
            summary: delete users
            description: Delete a user by username
            operationId: users#delete
            parameters:
                - name: username
                  in: path
                  description: Username of the user to delete
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
            schemes:
                - http
    /users/auth:
        post:
            tags:
                - users
            summary: auth users
            description: Authenticate a user with username and password
            operationId: users#auth
            parameters:
                - name: AuthRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/UsersAuthRequestBody'
                    required:
                        - username
                        - password
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/UsersAuthResponseBody'
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
            schemes:
                - http
definitions:
...
    User:
        title: User
        type: object
        properties:
            created_at:
                type: string
                description: User creation timestamp
                example: Qui odio minima consequatur et.
            updated_at:
                type: string
                description: User last update timestamp
                example: Qui et.
            username:
                type: string
                description: The username
                example: Aut hic maiores officia eum.
        example:
            created_at: Sint asperiores repudiandae deleniti sit rerum eveniet.
            updated_at: Tenetur maxime quasi veniam autem temporibus.
            username: Qui nulla hic molestias esse in.
        required:
            - username
            - created_at
            - updated_at
    UsersAuthRequestBody:
        title: UsersAuthRequestBody
        type: object
        properties:
            password:
                type: string
                description: Password
                example: s3cr3t
            username:
                type: string
                description: Username
                example: alice
        example:
            password: s3cr3t
            username: alice
        required:
            - username
            - password
    UsersAuthResponseBody:
        title: UsersAuthResponseBody
        type: object
        properties:
            token:
                type: string
                description: JWT or session token
                example: Laborum ratione minima.
            username:
                type: string
                description: The username of the authenticated user
                example: Cum ipsum tenetur blanditiis est pariatur.
        example:
            token: Sint ipsa debitis voluptatum non ducimus natus.
            username: Et quis vel reiciendis.
    UsersCreateRequestBody:
        title: UsersCreateRequestBody
        type: object
        properties:
            password:
                type: string
                description: Password (hashed or plain depending on implementation)
                example: s3cr3t
            username:
                type: string
                description: Username of the new user
                example: alice
        example:
            password: s3cr3t
            username: alice
        required:
            - username
            - password
    UsersCreateResponseBody:
        title: UsersCreateResponseBody
        type: object
        properties:
            username:
                type: string
                description: The username of the created user
                example: Error minus et maxime impedit necessitatibus sint.
        example:
            username: Perferendis est.